+++ examples/ex_rsrc/resource.h
@@ -1,0 +1,1 @@
+#define RCDATA1 101