+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.5.2 (Unreleased)
+* TBD
+
+# 1.5.1
+* Use dedicated symbolic expression attributes.
+
+# 1.5.0
+* Support Souffle interpreter for development.
+* Add support for loading existing GTIRB files.
+
 # 1.4.0
 
 * Add PE support.